requirements.txt
fastapi
uvicorn[standard]
openai
redis
python-dotenv
whisper
gTTS
pydantic
tiktoken
sentence-transformers
faiss-cpu
numpy
sounddevice
scipy
textblob
cryptography
better_profanity
websockets
requests
pydub
